#10 (#13)
made recomended sonarcloud fixes, added badges and added more generations to TrainnerTest to avoid bad results. may have to remove test in future

+++ src/main/java/com/lee/mcfaul/github/simple/ga/neural/network/Trainner.java
@@ -44,7 +44,7 @@ public class Trainner {
     NeuralNetwork highestScoringNetwork = null;
 
     double highestPossibleScore =
-        trainingSet.size() * trainingSet.get(0).get(1).length * (CORRECT_ELEMENT_VALUE
+        trainingSet.size() * trainingSet.get(0).get(1).length * (double) (CORRECT_ELEMENT_VALUE
             + PERFECT_ANSWER_VALUE);
 
     for (int i = 0; i < generations; i++) {
